update Convert to use VerbInput
