+++ javascript/react-controls/src/controls/DateFormatPatternCombobox.tsx
@@ -6,8 +6,8 @@ import type { IComboBoxProps } from '@fluentui/react'
 import { ComboBox, SelectableOptionMenuItemType } from '@fluentui/react'
 import { memo, useMemo } from 'react'
 
-import { dropdownStyles } from '../../styles.js'
-import { useDateFormatPatternOptions } from './DateFormatPatternCombobox.hooks.js'
+import { dropdownStyles } from '../styles.js'
+import { useDateFormatPatternOptions } from '../hooks.js'
 
 export interface DateFormatPatternComboboxProps
 	extends Partial<IComboBoxProps> {
