+++ javascript/react-controls/src/controls/index.ts
@@ -7,7 +7,7 @@ export * from './ColumnOrValueComboBox.js'
 export * from './ColumnSpread.js'
 export * from './ColumnValueDropdown.js'
 export * from './ContextualMenuItemSearchBox/index.js'
-export * from './DateFormatPatternCombobox/index.js'
+export * from './DateFormatPatternCombobox.js'
 export * from './EnumDropdown.js'
 export * from './GroupHeader/index.js'
 export * from './MultiDropdown.js'
