+++ javascript/react-controls/src/hooks.ts
@@ -61,3 +61,21 @@ function formatEnumName(name: string): string {
 	const rest = parts.slice(1).map(p => p.toLocaleLowerCase())
 	return [first, ...rest].join(' ')
 }
+
+export function useDateFormatPatternOptions(): IDropdownOption[] {
+	return useMemo(() => getDateFormatPatternOptions(), [])
+}
+
+export function getDateFormatPatternOptions(): IDropdownOption[] {
+	const formatPatternArray: IDropdownOption[] = [
+		{ key: '%Y-%m-%d', text: '%Y-%m-%d' },
+		{ key: '%Y/%m/%d', text: '%Y/%m/%d' },
+		{ key: '%B %d, %Y', text: '%B %d, %Y' },
+		{ key: '%m-%d-%Y', text: '%m-%d-%Y' },
+		{ key: '%m/%d/%Y', text: '%m/%d/%Y' },
+		{ key: '%d-%m-%Y', text: '%d-%m-%Y' },
+		{ key: '%d/%m/%Y', text: '%d/%m/%Y' },
+		{ key: '%Y-%m-%dT%H:%M:%S.%LZ', text: 'ISO 8601 (%Y-%m-%dT%H:%M:%S.%LZ)' },
+	]
+	return formatPatternArray
+}
